requirements.txt
dwave-ocean-sdk
pyqubo
matplotlib
seaborn
numpy
scipy